requirements.txt
google-genai==0.3.0
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-text-splitters==0.3.5
langchain-google-genai==2.0.8
langchain-chroma==0.2.0
chromadb==0.5.23
python-dotenv==1.0.1
pillow==11.1.0
pydantic==2.10.5
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20